+++ exporter/datadogexporter/config/config_test.go
@@ -101,18 +101,6 @@ func TestOverrideMetricsURL(t *testing.T) {
 	assert.Equal(t, cfg.Metrics.Endpoint, DebugEndpoint)
 }
 
-// TestDefaultSite tests that the Site option is set to the
-// default value when no value was set prior to running Sanitize
-func TestDefaultSite(t *testing.T) {
-	cfg := Config{
-		API: APIConfig{Key: "notnull"},
-	}
-
-	err := cfg.Sanitize(zap.NewNop())
-	require.NoError(t, err)
-	assert.Equal(t, cfg.API.Site, DefaultSite)
-}
-
 func TestValidate(t *testing.T) {
 
 	tests := []struct {
@@ -221,7 +209,7 @@ func TestUnmarshal(t *testing.T) {
 	tests := []struct {
 		name      string
 		configMap *config.Map
-		cfg       Config
+		cfg       *Config
 		err       string
 	}{
 		{